requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiogram==3.3.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
alembic==1.13.1 